Ccb allocation

+++ src/core/pmm.c
@@ -76,7 +76,7 @@ void pmm_init()
 	// buffers
 	pmm.entries = fmem = (OSZ_pmm_entry*)((uint8_t*)&__bss_start + 3*__PAGESIZE);
 		pmm.bss = (uint8_t*)&__bss_start;
-	pmm.bss_end = (uint8_t*)&__bss_start + __PAGESIZE * (nrphymax+3);
+	pmm.bss_end = (uint8_t*)&__bss_start + (uint64_t)(__PAGESIZE * ((uint)nrphymax+3));
 	// this is a chicken and egg scenario. We need free memory to
 	// store the free memory table...
 	while(num>0) {
